feat: 곱셈 기능 구현
