+++ src/main/java/Calculator/Multiplication.java
@@ -1,0 +1,7 @@
+package Calculator;
+
+public class Multiplication {
+    public int multiply(int firstNumber, int secondNumber) {
+        return firstNumber * secondNumber;
+    }
+}
